requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-templated-mail==1.1.1
djangorestframework==3.10.3
djoser==2.0.3
get==2019.4.13
idna==2.8
oauthlib==3.1.0
post==2019.4.13
public==2019.4.13
python3-openid==3.1.0
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
urllib3==1.25.7
